+++ project_Backend/vehicle_table.sql
@@ -8,9 +8,9 @@ CREATE TABLE vehicle_data(
     car_ID int NOT NULL AUTO_INCREMENT,
     car_make varchar(255),
     car_model varchar(255),
-    trunk_space float, 
-    foldable_seat_space float, 
-    personal_items float, 
+    trunk_space decimal, 
+    foldable_seat_space decimal, 
+    personal_items decimal, 
     
     PRIMARY KEY (car_ID)
 );
